Add flume integration with python generator, fix some bugs in generator, finished structure stream, refactor, add test json, add unit test

+++ spark-detect-bot-project/src/main/scala/com/dbocharov/detect/model/Models.scala
@@ -1,7 +1,7 @@
 package com.dbocharov.detect.model
 case class Event(
                   unix_time:Long,
-                  category_id:Int,
+                  category_id:Long,
                   ip:String,
                   event: String
                 )
@@ -9,3 +9,4 @@ case class BotRecord(
                       ip:String,
                       block_date:Long
                     )
+
